cleancode
# message auto-generated for no-merge-commit merge:
merge cherry-pick-mr-54-1740964125719-auto into master

merge "CleanCode" into master


Match-id-27a92dcb306cfce0728bc1c991c5c2483b9baaaa

+++ src/server/mirror_server.cpp
@@ -2321,9 +2321,10 @@ TraceDatabase MirrorServer::GetTraceData()
             traceData.metrics.goodEnd = traceInfo.GetGoodEnd();
             traceData.metrics.badEnd = traceInfo.GetBadEnd();
             traceData.metrics.min = (traceInfo.GetMin() == UINT64_MAX ?
-                                     0 : ((double)traceInfo.GetMin() / ock::htracer::unitStep));
-            traceData.metrics.max = static_cast<double>(traceInfo.GetMax()) / ock::htracer::unitStep;
-            traceData.metrics.total = static_cast<double>(traceInfo.GetTotal()) / ock::htracer::unitStep;
+                0 : static_cast<int>(static_cast<double>(traceInfo.GetMin()) / ock::htracer::unitStep));
+            traceData.metrics.max = static_cast<int>(static_cast<double>(traceInfo.GetMax()) / ock::htracer::unitStep);
+            traceData.metrics.total = static_cast<int>(static_cast<double>(traceInfo.GetTotal()) /
+                ock::htracer::unitStep);
 
             traceDatabase.traces[traceCount] = traceData;
             traceCount++;
